metadata and resolution routes
With this, the client may ask for the unresolved or resolved state for
any game parameterized on a turn, round pair, as well as the game
meatdata which includes, among other things, the latest turn and round
of the game.

An example use case: using the metadata, a client can decide whether to
show forward/backward buttons for browsing the history of a game, and
may use the /resolution route to grab information about what happened on
that turn.