HTTPS is optional; auto redirect (best effort)
HTTPS support should be dropped. To do TLS, put a proper web server in
front of the diplomacy server and proxy requests to it. But make sure
you have a domain name and a legit certificate... probably too much work
for a casual game.

Leaving it in because major browsers these days tend to try to force the
user into HTTPS, even when they're using intranet addresses
(192.168.0.0 or something). It can be overridden but it's a pain.

So this application of HTTP is in a weird spot: I like that everybody's
phone has a browser, so everybody's phone can play diplomacy on my local
area network. But browsers are (rightly so) moving towards HTTPS
everywhere; it's incresingly hard/awkward to opt-out, so I need a domain
name and a certificate. The "proper" solution for local area network
diplomacy seems to be:

  1. Get a domain name and some credentials for it (let's encrypt).
  2. Point that domain name to my local diplomacy server (not
     addressable from the internet), possibly via the DNS server in
     the local access point.
  3. Have players use that domain name.